Update Asset-Inventory.sql
The host function that was wrapping the da.ip_address in the group by throws an error in the report schema and doesn't validate. I removed that wrapper, and for syntax got rid of the quotes around dsite.name at the very beginning.

This validates in my lab console now and shows expected results.